Mobile update
This update adds a nicer looking mobile UI.

The navigation bar forms a 3-column display on devices with a width
of 60em or lower and a 2-column display on devices with a width of
45em or lower. It also becomes relatively positioned on devices 60em
or lower.

Animations are disabled on devices with a width of 60em or lower
because the window.onScroll method is only triggered on mobile devices
when the user releases their finger. Because of this, animations
look buggy on mobile devices.

People with pictures and links on the website form a tableview display
on devices with a width of 30em and lower. This makes it easy for people on
mobile devices to scroll through more content quicker and view the links
easier.

It might be useful if someone could figure out how to change the height of
the title banner on mobile devices.